Added a new file for depreacated functions. This can be removed later. Added basic crosshand_phase calculator.. but lots of work needed on that front. Minor changes done in other parts.
